поправил вьюху, делай prewiew

+++ app/models/place.rb
@@ -7,11 +7,14 @@ class Place < ActiveRecord::Base
   validates_attachment_content_type :original_image, :content_type => ["image/jpg", "image/jpeg", "image/png", "image/gif"]
 
   def crop_image
-    img = Magick::Image.read(self.image.path)[0]
-    img.crop!(crop_x.to_i, crop_y.to_i, crop_width.to_i, crop_height.to_i)
-    file = Tempfile.new(['image', '.png'])
-    img.write(self.image.path)
-    file.close
-    file.unlink
+    unless crop_x.blank?
+      img = Magick::Image.read(self.image.path)[0]
+      img.crop!(crop_x.to_i, crop_y.to_i, crop_width.to_i, crop_height.to_i, true)
+      file = Tempfile.new(['image', '.png'])
+      img.write(file.path)
+      self.update_attribute(:image, file)
+      file.close
+      file.unlink
+    end
   end
 end
